Move Header to components folder
