--- Frontend/src/features/shared/css/Header.css
@@ -1,7 +1,0 @@
-#header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-
-  padding-block: 1rem;
-}
