--- Frontend/src/features/shared/css/styles.css
@@ -1,1 +1,0 @@
-@import url('./Header.css');
