+++ Frontend/src/pages/Layout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router';
-import { Header } from '../features/shared/components';
+import Header from '../components/Header';
 
 const Layout = () => {
   return (
