+++ Frontend/src/styles/global.css
@@ -1,3 +1,11 @@
 .g-container {
   padding-inline: 4rem;
 }
+
+#header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+
+  padding-block: 1rem;
+}
