+++ Frontend/src/styles/index.css
@@ -2,4 +2,3 @@
 @import url('./global.css');
 @import url('../features/auth/css/styles.css');
 @import url('../features/companies/css/styles.css');
-@import url('../features/shared/css/styles.css');
